old-tests/submission/W3C/domhtml/HTMLImageElement04.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>NIST DOM HTML Test - IMG</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/domtests-common.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-img-border" rel="help">
<link rel="author" title="NIST">
</head>
<body>
<p>
<img align="middle" alt="DTS IMAGE LOGO" border="0" height="50" hspace="4" id="IMAGE-1" longdesc="./files/desc.html" name="IMAGE-1" src="/support/images/yellow-rectangle.png" usemap="#DTS-MAP" vspace="10" width="100">
</p>



<script>
var domtest = async_test("The name, align, border, hspace, and vspace IDL attributes of the img element must reflect the respective content attributes of the same name.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var vborder;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("img");
      assert_equals(nodeList.length,1);
testNode = nodeList.item(0);
      vborder = testNode.border;

      assert_equals(vborder,"0");
       
    domtest.done();
}));

</script>
<p>The name, align, border, hspace, and vspace IDL attributes of the img element must reflect the respective content attributes of the same name.</p>
<div id="log"></div>
</body></html>
